Change reports file to TS, add types
